backend/requirements.txt
# Existing packages
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.5.0
python-dotenv==1.0.0
groq==0.4.0


# NEW: Tiered extraction dependencies
trafilatura==1.6.0          # Tier 1: Reader mode content extraction
playwright==1.40.0          # Tier 2: Headless browser (run: playwright install)
aiohttp==3.9.0              # Async HTTP for concurrent fallbacks
